Fix identity db tool (#47)
* reworked db tool

* Corrected copyright
