--- scripts/dev/buildAccountsDB.sh
@@ -1,3 +1,0 @@
-export PYTHONPATH=items/tool_accounts_db_builder:items/shared
-
-python items/tool_accounts_db_builder/db_builder.py
