+++ scripts/dev/buildIdentityDB.sh
@@ -1,0 +1,3 @@
+export PYTHONPATH=items/tool_identity_db_builder:items/shared
+
+python items/tool_identity_db_builder/db_builder.py
